doc: compile tutorial programs
The programs in `doc/examples` are not compiled currently, so let's
compile them if the `docs` and `examples` options don't disallow it.

`tutorial4.c` needs a small modification to avoid `-Wfloat-conversion`.

Additionally, install them if `installed_tests` is not disabled.

+++ doc/examples/tutorial4.c
@@ -56,7 +56,7 @@ static void on_process(void *userdata)
 		 * Another common method to convert a double to
 		 * 16 bits is to multiple by 32768.0 and then clamp to
 		 * [-32768 32767] to get the full 16 bits range. */
-		val = sin(data->accumulator) * DEFAULT_VOLUME * 32767.0;
+		val = (int16_t) (sin(data->accumulator) * DEFAULT_VOLUME * 32767.0);
 		for (c = 0; c < DEFAULT_CHANNELS; c++)
 			*dst++ = val;
 	}
